# Changed HashMap.getOrElseUpdate to only calculate the index once