@paplorinc 2016-11-14
Signed the Scala CLA