@retronym 2016-11-15
The benchmark could be added to our (recently added) JMH suite, in the same place as [OpenHashMapBenchmark](https://github.com/scala/scala/blob/100d6374c282d94497ee9f0d4a206d427951c74c/test/benchmarks/src/main/scala/scala/collection/mutable/OpenHashMapBenchmark.scala).

We don't run these automatically, but it will at least provide a reference for someone studying this commit or making further changes to this area.

The benchmark itself could do with tests of workloads that exercise more code paths. For instance, using different key types will lead control through through more paths in `BoxesRuntime.{equals*, hash*}`, and different function arguments to produce the default to make the call to `Function.apply` within `getOrElseUpdate` megamorphic.