@Ichoran 2016-11-15
@paplorinc - As an aside: I'm not sure why using JMH from Scala rather than Java is "counter-intuitive" for you.  A benchmark that depends on what language you use to embed it is almost by definition a bad benchmark because it means that the incidental stuff that you are not supposed to care about is not negligibly inexpensive compared to what you're actually interested in.
If there is a difference, I still wouldn't expect it to necessarily work out in favor of using Java for Scala (and vice-versa, I suppose): by removing the Scala code from a more normal context, you might make the microbenchmarking results even less relevant to normal usage than microbenchmarks usually are.  For instance, if javac uses a pattern around the benchmark that is really easy for the JIT compiler to optimize away, but almost all use cases are from Scala with scalac's not-exactly-the-same pattern, you might fail to see an effect that is bothering everyone.
But really, the point is if it matters, it's already being done wrong.  (The ops/s reported here are pretty well into the safe range where you could even call with interpreted Groovy and it'd be fine.)