@Ichoran 2016-11-15
I think the core question that I don't have an answer to is whether we're allowed to introduce an override to getOrElseUpdate and still stay binary compatible.  If yes, then I think the other issues can be worked around by duplicating code (inlining it manually) until you end up with a set of calls to protected or public methods.  If yes in both 2.12 and 2.11, maybe this patch could be backported!  Faster hash maps benefit almost everyone.