@retronym 2016-11-16
> I think the core question that I don't have an answer to is whether we're allowed to introduce an override to getOrElseUpdate and still stay binary compatible

This is okay because the class already inherits a method with and identical signature from `AbstractMap`.  Client code of 2.12.1 may include `invokevirtual HashMap.getOrElseUpdate`, but this will not incur an `IncompatibleClassChangeError` if linked against the 2.12.0 class thanks to the override.

Here's an example of making this class change with Java code without linkage errors: https://gist.github.com/794c9f88771b210f5dab0b65793b4682