@Ichoran 2016-11-16
@retronym - Okay, then I _think_ we're clear once @paplorinc switches to use `table` and friends instead of `addEntry0` etc..  Hopefully that will still show a speedup.  The JIT compiler won't be able to take advantage of its efforts in already optimizing `addEntry0`, but hopefully it'll do just as well on the inlined version.