@paplorinc 2016-11-16
@Ichoran, to be sure I understand what you're requesting, do you want me to keep copy-pasting things from `HashTable` until everything compiles? That would result in duplicating even some fields:
```scala
override def getOrElseUpdate(key: A, defaultValue: => B): B = {
  val i = index(elemHashCode(key))
  val entry = findEntry0(key, i)
  if (entry != null)
    entry.value
  else {
    val newEntry = createNewEntry(key, defaultValue)
    addEntry0(newEntry, i)
    newEntry.value
  }
}
private[this] final def findEntry0(key: A, h: Int): Entry = {
  var e = table(h).asInstanceOf[Entry]
  while (e != null && !elemEquals(e.key, key)) e = e.next
  e
}
private[this] final def addEntry0(e: Entry, h: Int) {
  e.next = table(h).asInstanceOf[Entry]
  table(h) = e
  tableSize = tableSize + 1
  nnSizeMapAdd(h)
  if (tableSize > threshold)
    resize(2 * table.length)
}
private def resize(newSize: Int) {
  val oldTable = table
  table = new Array(newSize)
  nnSizeMapReset(table.length)
  var i = oldTable.length - 1
  while (i >= 0) {
    var e = oldTable(i)
    while (e != null) {
      val h = index(elemHashCode(e.key))
      val e1 = e.next
      e.next = table(h).asInstanceOf[Entry]
      table(h) = e
      e = e1
      nnSizeMapAdd(h)
    }
    i = i - 1
  }
  threshold = newThreshold(_loadFactor, newSize)
}
private[collection] final def newThreshold(_loadFactor: Int, size: Int) = ((size.toLong * _loadFactor) / loadFactorDenom).toInt
private[collection] final def loadFactorDenom = 1000
```