@Ichoran 2016-11-16
@paplorinc - Which fields do you need?  You've got everything already there as protected, don't you?

It might be easier to intercept the size and just eat the extra lookup on an addEntry in the relatively rare case that the table needs to be enlarged.

I grant that it's ugly either way, but if we can avoid even a minor binary incompatibility, I think it's for the best.  If we'd caught this before 2.12 was out, I'd have much preferred your patch as it was where you just change the visibility of some things.

Also, I understand your point in general regarding using a flawed system to detect its own flaws, but in this particular case, do you also understand my point?