@retronym 2016-11-16
The binary incompatibility wasn't reported because of a regression in the entry point into MiMa we use. Will be fixed tomorrow: https://github.com/typesafehub/migration-manager/pull/138