@paplorinc 2016-11-16
@Ichoran, thanks for the idea, I've tested it from the new `Scala` benchmarks and they maintained their speed advantage (except for the case when it actually needs growing, but that's ok), and the code is also still maintainable - great balance!

Please wait with the merge until I validate the results from `Java` also (I get some strange results that I need to investigate)

@retronym, thank you for your investigations!