@Ichoran 2016-11-16
Looks like you accidentally got a couple of stray commits in here?