@paplorinc 2016-11-18
@Ichoran, not sure why the build is failing do I still have a `MiMa` violation?
[`library/mima failed: java.lang.RuntimeException: MiMa failed with exit code 1`](https://scala-ci.typesafe.com/job/scala-2.12.x-validate-test/3547/console)