@szeiger 2016-11-18
```
[info] Checking backward binary compatibility
[info] prev = /home/jenkins/.ivy2/cache/org.scala-lang/scala-library/jars/scala-library-2.12.0.jar, curr = /home/jenkins/workspace/scala-2.12.x-validate-test@2/build/pack/lib/scala-library.jar
Found 1 binary incompatibilities
================================
 * method getExitValue()scala.Tuple2 in class
   scala.sys.process.ProcessImpl#CompoundProcess does not have a correspondent
   in current version

Generated backward filter config definition
============================================

    filter {
        problems=[
            {
                matchName="scala.sys.process.ProcessImpl#CompoundProcess.getExitValue"
                problemName=DirectMissingMethodProblem
            }
        ]
    }


Generated forward filter config definition
===========================================

    filter {
        problems=[]
    }
```