@szeiger 2016-11-18
This looks like an unrelated failure due to MiMa having been accidentally disabled for a while.